refactor: create contents move to post router
