+++ src/api/routes/post.router.ts
@@ -7,6 +7,8 @@ import { projectPrisma } from '../database/prisma/client.js';
 import { asyncHandler } from '../middlewares/asyncHandler.js';
 import { authenticatedUser } from '../middlewares/auth.js';
 import { validateAccess } from '../middlewares/validateAccess.js';
+import { createContentUseCaseSchema } from '../useCases/content/createContent.schema.js';
+import { CreateContentUseCase } from '../useCases/content/createContent.useCase.js';
 import { changeStatusUseCaseSchema } from '../useCases/post/changeStatus.schema.js';
 import { ChangeStatusUseCase } from '../useCases/post/changeStatus.useCase.js';
 import { createPostUseCaseSchema } from '../useCases/post/createPost.schema.js';
@@ -89,6 +91,32 @@ router.post(
 
     res.json({
       post,
+    });
+  })
+);
+
+router.post(
+  '/posts/:id/contents',
+  authenticatedUser,
+  validateAccess(['createPost']),
+  asyncHandler(async (req: Request, res: Response) => {
+    const validated = createContentUseCaseSchema.safeParse({
+      projectId: res.projectRole?.id,
+      id: req.params.id,
+      locale: req.body.locale,
+      userId: res.user.id,
+    });
+    if (!validated.success) throw new InvalidPayloadException('bad_request', validated.error);
+
+    const useCase = new CreateContentUseCase(
+      projectPrisma(validated.data.projectId),
+      new ContentRepository(),
+      new ContentHistoryRepository()
+    );
+    const content = await useCase.execute(validated.data);
+
+    res.json({
+      content,
     });
   })
 );
